project/requirements.txt
fastapi
uvicorn
httpx
pandas
scikit-learn
psycopg2-binary==2.9.10
streamlit
sentence_transformers
nltk
datasets
